requirements.txt
aiogram==3.3.0
aiohttp
python-dotenv